Dev/4.5.0

+++ .github/ci/build/build_android.sh
@@ -49,27 +49,33 @@ echo release_version: $release_version
 echo short_version: $short_version
 echo pwd: `pwd`
 echo sdk_url: $sdk_url
+unzip_name=Agora_Native_SDK_for_Android_FULL_DEFAULT
+zip_name=Agora_Native_SDK_for_Android_FULL_DEFAULT.zip
+if [ -z "$sdk_url" ]; then
+   echo "sdk_url is empty"
+   echo unzip_name: $unzip_name 
+   echo zip_name: $zip_name
+else
+	zip_name=${sdk_url##*/}
+	echo zip_name: $zip_name
 
-zip_name=${sdk_url##*/}
-echo zip_name: $zip_name
+	# env LC_ALL=en_US.UTF-8 python3 $WORKSPACE/artifactory_utils.py --action=download_file --file=$sdk_url || exit 1
+	curl -o $zip_name $sdk_url || exit 1
+	7za x ./$zip_name -y > log.txt
 
-# env LC_ALL=en_US.UTF-8 python3 $WORKSPACE/artifactory_utils.py --action=download_file --file=$sdk_url || exit 1
-curl -o $zip_name $sdk_url || exit 1
-7za x ./$zip_name -y > log.txt
+	unzip_name=`ls -S -d */ | grep Agora | sed 's/\///g'`
+	echo unzip_name: $unzip_name
 
-unzip_name=`ls -S -d */ | grep Agora | sed 's/\///g'`
-echo unzip_name: $unzip_name
-
-rm -rf ./$unzip_name/rtc/bin
-rm -rf ./$unzip_name/rtc/demo
-rm ./$unzip_name/rtc/commits
-rm ./$unzip_name/rtc/package_size_report.txt
-mkdir ./$unzip_name/rtc/samples
-
+	rm -rf ./$unzip_name/rtc/bin
+	rm -rf ./$unzip_name/rtc/demo
+	rm ./$unzip_name/rtc/commits
+	rm ./$unzip_name/rtc/package_size_report.txt
+	rm -rf ./$unzip_name/pom
+fi
+mkdir -p ./$unzip_name/rtc/samples 
 cp -rf ./Android/${android_direction} ./$unzip_name/rtc/samples/API-Example || exit 1
 7za a -tzip result.zip -r $unzip_name > log.txt
 mv result.zip $WORKSPACE/withAPIExample_${BUILD_NUMBER}_$zip_name
-
 
 if [ $compile_project = true ]; then
 	# install android sdk
@@ -79,7 +85,12 @@ if [ $compile_project = true ]; then
 	export ANDROID_HOME=/usr/lib/android_sdk
 	echo ANDROID_HOME: $ANDROID_HOME
 	cd ./$unzip_name/rtc/samples/API-Example || exit 1
-	./cloud_build.sh || exit 1
+	if [ -z "$sdk_url" ]; then
+		./cloud_build.sh false || exit 1
+	else
+		./cloud_build.sh true || exit 1
+	fi
+	
 fi
